Change scene, fixed planes, ...

+++ CudaRaytracer/color.h
@@ -11,7 +11,7 @@ struct Color
 	__device__ Color(float const& ared, float const& agreen, float const& ablue)
 		: red(ared), green(agreen), blue(ablue)
 	{ }
-	__device__ void accumulate(Color const& x, double const& scale = 1.0) 
+	__device__ void accumulate(Color const& x, float const& scale = 1.0) 
 	{ red += scale*x.red; green += scale*x.green; blue += scale*x.blue; }
 
 	__device__ Color &operator *= (float const& x) { red *= x; green *= x; blue *= x; return *this; }
